Move formatters together
